clean up: remove CPU treeverify path
